# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FlowVerse 3D HTTP API
  version: 0.1.0
  summary: API para diseño, versionado, validación y simulación de flujos.
  description: |
    Contrato HTTP canónico del MVP. La autenticación utiliza cookies HttpOnly
    opacas. Toda mutación autenticada debe incluir `X-CSRF-Token`. Los errores
    comparten un sobre estable y cada respuesta incluye `X-Request-Id`.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
security:
  - cookieAuth: []
paths:
...
  /v1/flows/{flowId}/diff:
    parameters:
      - $ref: "#/components/parameters/FlowId"
    get:
      tags: [Versions]
      operationId: diffFlowRevisions
      summary: Compara semánticamente dos revisiones de un flujo.
      description: |
        Un selector omitido representa el borrador actual. Debe indicarse al
        menos una versión; esto permite versión→draft, draft→versión o
        versión→versión. El orden de nodos, conexiones y variables no genera
        ruido, y cada cambio se clasifica como visual, behavioral o breaking.
      parameters:
        - name: baseVersionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: targetVersionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Diff determinista entre las dos revisiones.
          content:
            application/json:
              schema:
                $ref: "./schemas/flow-diff.schema.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowverse_access
...
  parameters:
...
    FlowId:
      name: flowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
    Issue:
      type: object
      additionalProperties: false
      required: [code, severity, message]
      properties:
        code:
          type: string
        severity:
          enum: [info, warning, error]
        message:
          type: string
        path:
          type: string
        nodeId:
          type: string
        edgeId:
          type: string
...
  responses:
...
    BadRequest:
      description: |
        Solicitud inválida. En creación de runs incluye
        `idempotency.key_required`, `idempotency.key_invalid`, JSON inválido u
        overrides inexistentes.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Recurso inexistente o no visible para el usuario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: Límite temporal excedido.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...